requirements.txt
amqp==5.3.1
anyio==4.7.0
async-timeout==5.0.1
billiard==4.2.1
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
fastapi==0.95.2
h11==0.14.0
idna==3.10
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.2
motor==3.6.0
numpy==2.1.3
pandas==2.2.3
prompt_toolkit==3.0.48
pydantic==1.10.2
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
redis==5.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.22.0
vine==5.1.0
wcwidth==0.2.13